Deprecate `pyramid.compat`.
