+++ .flake8
@@ -10,6 +10,7 @@ ignore =
     W504
 exclude =
     src/pyramid/compat.py
+    src/pyramid/_compat.py
     src/pyramid/scaffolds/alchemy
     src/pyramid/scaffolds/starter
     src/pyramid/scaffolds/zodb
